input & inputUI 분리
