+++ src/main/java/input/InputUI.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Scanner;
 import lotto.Lotto;
 
-public class Input {
+public class InputUI {
     private List<Lotto> lottos = new ArrayList<>();
     private Scanner scanner = new Scanner(System.in);
     private List<String> right;
@@ -24,22 +24,22 @@ public class Input {
         System.out.println("---------");
     }
 
-    public List<Lotto> makeLottos(int num) {
-        for(int i = 0; i < num; i++) {
-            Lotto lotto = new Lotto();
-            lottos.add(lotto);
-            System.out.println(lotto.getLotto());
-        }
-        return lottos;
-    }
-
-    public List<Lotto> addManualLottos(List<String> manual) {
-        for(int i = 0; i < manual.size(); i++) {
-            Lotto lotto = new Lotto(manual.get(i));
-            lottos.add(lotto);
-        }
-        return lottos;
-    }
+//    public List<Lotto> makeLottos(int num) {
+//        for(int i = 0; i < num; i++) {
+//            Lotto lotto = new Lotto();
+//            lottos.add(lotto);
+//            System.out.println(lotto.getLotto());
+//        }
+//        return lottos;
+//    }
+//
+//    public List<Lotto> addManualLottos(List<String> manual) {
+//        for(int i = 0; i < manual.size(); i++) {
+//            Lotto lotto = new Lotto(manual.get(i));
+//            lottos.add(lotto);
+//        }
+//        return lottos;
+//    }
 
     public String right() {
         String input = scanner.nextLine();
