+++ src/main/java/lotto/Lotto.java
@@ -1,6 +1,5 @@
 package lotto;
 
-import input.Input;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
