+++ src/main/java/lotto/LottoGame.java
@@ -1,0 +1,14 @@
+//package lotto;
+//
+//import input.InputCommon;
+//
+//import java.util.List;
+//
+//public class LottoGame {
+//    private List<Lotto> lottos;
+//    LottoGame(int inputMoney, List<String> manualNum) {
+//        InputCommon input = new InputCommon();
+//        input.makeLottos(inputMoney);
+//        input.addManualLottos(manualNum);
+//    }
+//}
