+++ src/main/java/rank/Check.java
@@ -1,20 +1,20 @@
 package rank;
 
-import input.Input;
+import input.InputUI;
 import lotto.Lotto;
 
 import java.util.List;
 
 public class Check {
-    Input input = new Input();
-    private Lotto winningNumber = new Lotto(input.right());
+    InputUI inputUI = new InputUI();
+    private Lotto winningNumber = new Lotto(inputUI.right());
 
     public Lotto getRight() {
         return winningNumber;
     }
 
     public void checking(List<Lotto> lottos) {
-        Input.bonusNumber();
+        InputUI.bonusNumber();
         for(int i = 0; i < lottos.size(); i++) {
             lottos.get(i).checkNumber(winningNumber);
         }
