+++ src/main/resources/templates/index.html
@@ -58,9 +58,9 @@ body {
 	  <div class="col-md-4">
 		<form class="form-lotto" method="post" action="/buyLotto">
 			<h2 class="text-center">로또 게임</h2>
-			<div class="input-group">
-				<div class="input-group-addon">￦</div>
-				<input type="number" id="inputMoney" class="form-control"
+			<div class="inputUI-group">
+				<div class="inputUI-group-addon">￦</div>
+				<inputUI type="number" id="inputMoney" class="form-control"
 					name="inputMoney" placeholder="금액 입력" required autofocus>
 			</div>
 			<div class="form-show-div" class="form-group">
