+++ src/main/resources/templates/show.html
@@ -65,12 +65,12 @@
 <form class="form-show" method="post" action="/matchLotto">
     <div class="form-show-div form-group">
         <label for="winningNumber">지난 주 당첨 번호</label>
-        <input type="text" id="inputMoney" class="form-control"
+        <inputUI type="text" id="inputMoney" class="form-control"
                name="winningNumber" placeholder="1,2,3,4,5,6" required autofocus>
     </div>
     <div class="form-show-div form-group">
         <label for="bonusNumber">2등 보너스 볼</label>
-        <input type="text" class="form-control" id="bonusNumber"
+        <inputUI type="text" class="form-control" id="bonusNumber"
                name="bonusNumber" required autofocus>
     </div>
     <div class="submit-button">
